Moved load_tron_yaml to utils to avoid circular imports
