+++ paasta_tools/tron_tools.py
@@ -39,6 +39,7 @@ from paasta_tools.utils import load_v2_deployments_json
 from paasta_tools.utils import NoConfigurationForServiceError
 from paasta_tools.utils import NoDeploymentsAvailable
 from paasta_tools.utils import paasta_print
+from paasta_tools.utils import load_tron_yaml
 
 from paasta_tools import monitoring_tools
 from paasta_tools.monitoring_tools import list_teams
@@ -524,17 +525,6 @@ def load_tron_instance_config(
                 if action.get_action_name() == requested_action:
                     return action
     raise NoConfigurationForServiceError(f"No tron configuration found for {service} {instance}")
-
-
-def load_tron_yaml(service: str, cluster: str, soa_dir: str) -> Dict[str, Any]:
-    config = service_configuration_lib.read_extra_service_information(
-        service_name=service,
-        extra_info=f'tron-{cluster}',
-        soa_dir=soa_dir,
-    )
-    if not config:
-        raise NoConfigurationForServiceError('No Tron configuration found for service %s' % service)
-    return config
 
 
 def extract_jobs_from_tron_yaml(config):
